pyproject.toml
[tool.poetry]
name = "genezio"
version = "0.1.0"
description = ""
authors = ["Alex Vesa & Rares Istoc"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pypdf = "^4.2.0"
openai = "^1.30.4"
python-dotenv = "^1.0.1"
pydantic = "^2.7.2"
qdrant-client = "^1.9.1"
fastembed = "^0.3.3"
streamlit = "^1.41.1"
llama-index = "^0.10.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
